Labbar/requirements.txt
streamlit
scikit-learn
pandas
numpy
Pillow
requests
